+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### BREAKING
 * Remove `railway_deployment_trigger` resource
+* Remove `region` from `railway_service` in favor of `regions` multi region support
 
 ### Bug fixes
 * Fixes issue with tcp proxy not being used in service after creating